Set halo <-> fof group pointers during read.

+++ src/core/reionization.c
@@ -311,7 +311,7 @@ int map_galaxies_to_slabs(int ngals)
     }
 
     // sort the slab indices IN PLACE (n.b. compare_slab_assign is a stable comparison)
-    qsort(galaxy_to_slab_map, gal_counter, sizeof(gal_to_slab_t), compare_slab_assign);
+    qsort(galaxy_to_slab_map, (size_t)gal_counter, sizeof(gal_to_slab_t), compare_slab_assign);
 
     assert(gal_counter == ngals);
